# frontend-mobile/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resultados River API Documentation",
        "description": "API for managing and querying historical football results for River Plate.",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/v1/estadios": {
            "get": {
                "tags": [
                    "Estadios"
                ],
                "summary": "List all estadios",
                "description": "Display a listing of the resource.",
                "operationId": "getEstadios",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Estadios"
                ],
                "summary": "Create a new estadio",
                "description": "Store a newly created resource in storage.",
                "operationId": "createEstadio",
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/estadios/{id}": {
            "get": {
                "tags": [
                    "Estadios"
                ],
                "summary": "Get estadio by ID",
                "description": "Display the specified resource.",
                "operationId": "getEstadioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Estadios"
                ],
                "summary": "Update an estadio",
                "description": "Update the specified resource in storage.",
                "operationId": "updateEstadio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Estadios"
                ],
                "summary": "Delete an estadio",
                "description": "Remove the specified resource from storage.",
                "operationId": "deleteEstadio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/fases": {
            "get": {
                "tags": [
                    "Fases"
                ],
                "summary": "List all fases",
                "description": "Display a listing of the resource.",
                "operationId": "getFases",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fases"
                ],
                "summary": "Create a new fase",
                "description": "Store a newly created resource in storage.",
                "operationId": "createFase",
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/fases/{id}": {
            "get": {
                "tags": [
                    "Fases"
                ],
                "summary": "Get fase by ID",
                "description": "Display the specified resource.",
                "operationId": "getFaseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Fases"
                ],
                "summary": "Update a fase",
                "description": "Update the specified resource in storage.",
                "operationId": "updateFase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fases"
                ],
                "summary": "Delete a fase",
                "description": "Remove the specified resource from storage.",
                "operationId": "deleteFase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/jugadores": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "List all jugadores",
                "description": "Display a listing of the resource.",
                "operationId": "getJugadores",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Create a new jugador",
                "description": "Store a newly created resource in storage.",
                "operationId": "createJugador",
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/jugadores/{id}": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Get jugador by ID",
                "description": "Display the specified resource.",
                "operationId": "getJugadorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Update a jugador",
                "description": "Update the specified resource in storage.",
                "operationId": "updateJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Delete a jugador",
                "description": "Remove the specified resource from storage.",
                "operationId": "deleteJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Sanctum token to authenticate",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}